+++ CHANGELOG.rst
@@ -26,6 +26,7 @@ Changelog
 - Add ``range_right``. Thanks bharadwajyarlagadda_!
 - Add ``subtract``. Thanks bharadwajyarlagadda_!
 - Add ``stub_array``. Thanks bharadwajyarlagadda_!
+- Add ``stub_dict``. Thanks bharadwajyarlagadda_!
 - Add ``stub_false``. Thanks bharadwajyarlagadda_!
 - Add ``stub_string``. Thanks bharadwajyarlagadda_!
 - Add ``stub_true``. Thanks bharadwajyarlagadda_!